Chore: Extended PoseSnapshot from 16 to 32 joints

+++ Assets/Arteranos/Scripts/Avatar/NetworkPose.cs
@@ -140,7 +140,7 @@ namespace Arteranos.NetworkIO
                 // Debug.Log($"Skipped snapshot queue for {name} to snapshot[{serverSnapshots.Count-1}]");
             }
 
-            ushort mask;
+            uint mask;
             if(initialState || lastSerializedRotations == null)
             {
                 // initial - serialize all of these
